aggregator/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
pandas>=2.0.0
scikit-learn>=1.2.2
matplotlib>=3.7.1
seaborn>=0.12.2
pillow>=9.5.0
tqdm>=4.65.0
mlflow>=2.3.0
python-multipart>=0.0.6     